Try tricky's suggestion to use three channels
This is louder and I don't hear any significant glitching in BeebEm or
jsbeeb. It does burn more CPU time though, which isn't ideal if I'm
playing under interrupts.

There was a bug which happened to be harmless in the code before;
writing the high bits of the tone should *not* include the channel
number. I got away with this for channel 0 but when I copied the code to
initialise channels 1 and 2 it went wrong.